Began general solver
Began setting up the framework for a solver of general electrostatics problems.
The 'board' is a 100x100 array of cells. The user inputs charged bodies via a text file which contains the dimensions and potential of any number of charged shapes. 

TODO:
Implement the relaxation stage so that the program actually does something useful.
Implement graphical output.
Add support for other types of shapes than rectangles (i.e. circles).